requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
aiosqlite==0.20.0
gspread==5.12.2
authlib==1.3.0
apscheduler==3.10.4
httpx==0.27.2
